refactor: extract GoogleSignInButton into separate reusable component
