+++ src/components/GoogleSignInButton.jsx
@@ -1,0 +1,91 @@
+// src/components/GoogleSignInButton.jsx
+import React from "react";
+import styled from "@emotion/styled";
+
+const GoogleButton = styled.button`
+  width: 100%;
+  padding: 14px 24px;
+  background: ${({ theme }) => theme.colors.surface};
+  color: ${({ theme }) => theme.colors.textPrimary};
+  border: 1px solid ${({ theme }) => theme.colors.border};
+  border-radius: 8px;
+  font-size: 15px;
+  font-weight: 500;
+  cursor: pointer;
+  transition: all 0.2s ease;
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  gap: 10px;
+  min-height: 48px;
+
+  &:hover {
+    background: ${({ theme }) => theme.colors.gray100};
+    border-color: ${({ theme }) => theme.colors.gray300};
+    transform: translateY(-1px);
+  }
+
+  &:active {
+    transform: translateY(0);
+  }
+
+  &:disabled {
+    background: ${({ theme }) => theme.colors.gray100};
+    color: ${({ theme }) => theme.colors.gray400};
+    cursor: not-allowed;
+    transform: none;
+  }
+
+  ${props =>
+    props.isLoading &&
+    `
+    background: ${({ theme }) => theme.colors.gray100};
+    color: ${({ theme }) => theme.colors.gray400};
+    cursor: not-allowed;
+  `}
+
+  @media (min-width: 480px) {
+    padding: 12px 24px;
+    font-size: 16px;
+    gap: 12px;
+  }
+`;
+
+// Google Icon Component
+const GoogleIcon = () => (
+  <svg width="20" height="20" viewBox="0 0 24 24" fill="none">
+    <path
+      fill="#4285F4"
+      d="M22.56 12.25c0-.78-.07-1.53-.2-2.25H12v4.26h5.92c-.26 1.37-1.04 2.53-2.21 3.31v2.77h3.57c2.08-1.92 3.28-4.74 3.28-8.09z"
+    />
+    <path
+      fill="#34A853"
+      d="M12 23c2.97 0 5.46-.98 7.28-2.66l-3.57-2.77c-.98.66-2.23 1.06-3.71 1.06-2.86 0-5.29-1.93-6.16-4.53H2.18v2.84C3.99 20.53 7.7 23 12 23z"
+    />
+    <path
+      fill="#FBBC05"
+      d="M5.84 14.09c-.22-.66-.35-1.36-.35-2.09s.13-1.43.35-2.09V7.07H2.18C1.43 8.55 1 10.22 1 12s.43 3.45 1.18 4.93l2.85-2.22.81-.62z"
+    />
+    <path
+      fill="#EA4335"
+      d="M12 5.38c1.62 0 3.06.56 4.21 1.64l3.15-3.15C17.45 2.09 14.97 1 12 1 7.7 1 3.99 3.47 2.18 7.07l3.66 2.84c.87-2.6 3.3-4.53 6.16-4.53z"
+    />
+  </svg>
+);
+
+const GoogleSignInButton = ({ onClick, isLoading, disabled, ...props }) => {
+  return (
+    <GoogleButton
+      type="button"
+      onClick={onClick}
+      disabled={disabled || isLoading}
+      isLoading={isLoading}
+      {...props}
+    >
+      <GoogleIcon />
+      {isLoading ? "Signing in..." : "Continue with Google"}
+    </GoogleButton>
+  );
+};
+
+export default GoogleSignInButton;
